Fix solver defects surfaced by MATLAB verification testing
All five fixes were found by comparing the solvers against the MATLAB
reference data in PMKS-Web/PMKS_Verification (tests added separately):

- ForceSolver added a parallel-axis m*d^2 term (d measured from the link
  CoM to an arbitrary loop joint) to the moment-of-inertia while taking
  moments about the CoM, skewing every dynamic force/torque result. The
  inertial term about the CoM is J*alpha alone.
- circleCircleIntersection rejected internally tangent circles, which
  arise whenever a joint is collinear with the two joints it is solved
  from (e.g. a tracer point on a straight ternary link). Positions are
  rounded every timestep, so exact tangency degraded into "no
  intersection" and invalidated the whole mechanism.
- PositionSolver never marked slider-connected joints as known and never
  recursed past them, so a tracer point on the slider's coupler link
  could not be ordered and the mechanism was rejected.
- KinematicsSolver looked up a piston's revolute joint by object identity
  in the per-timestep joint arrays, which hold copies; the index came
  back -1 and slider analysis crashed. It also never reset its cached
  input-joint index, leaking state between mechanisms.
- The joint -> coord -> settings -> utils -> joint import cycle broke
  class initialization when the model modules were entered from a second
  test entry point. The unit enums now live in their own module
  (re-exported from utils.ts) and joint.ts's Link import is type-only.
