+++ src/app/model/joint.ts
@@ -1,5 +1,8 @@
 import { AppConstants } from './app-constants';
-import { Link, RealLink } from './link';
+// Type-only: a runtime import here closes the joint -> link -> joint module
+// cycle, which breaks class initialization when a test entry point loads the
+// model modules in a different order than the app does.
+import type { Link } from './link';
 import { Coord } from './coord';
 import { SettingsService } from '../services/settings.service';
