+++ src/app/model/mechanism/kinematic-solver.ts
@@ -32,10 +32,13 @@ export class KinematicsSolver {
   private static groundJointIndexMap = new Map<string, number>();
   static realJointIndexMap = new Map<string, number>();
   static desiredAngleMap = new Map<string, number>();
-  private static inputJointIndex: number;
+  private static inputJointIndex: number | undefined;
   static inputLinkIndex: number;
 
   static resetVariables() {
+    // Also forget the cached input joint so the next call re-derives it (and
+    // inputLinkIndex) for the new mechanism instead of reusing stale indices.
+    this.inputJointIndex = undefined;
     this.jointIndexMap = new Map<string, number>();
     this.jointVelMap = new Map<string, [number, number]>();
     this.jointAccMap = new Map<string, [number, number]>();
@@ -140,7 +143,12 @@ export class KinematicsSolver {
           if (realJoint === undefined) {
             return;
           }
-          this.realJointIndexMap.set(links[this.inputLinkIndex].id, joints.indexOf(realJoint));
+          // Match by id: the per-timestep joint arrays hold copies, so an
+          // identity indexOf against the link's original joints finds nothing.
+          this.realJointIndexMap.set(
+            links[this.inputLinkIndex].id,
+            joints.findIndex((j) => j.id === realJoint.id)
+          );
           const prisJoint = links[this.inputLinkIndex].joints.find(jt => jt instanceof PrisJoint) as PrisJoint;
           this.desiredAngleMap.set(links[this.inputLinkIndex].id, prisJoint.angle_rad);
         }
@@ -215,7 +223,13 @@ export class KinematicsSolver {
             case Piston:
               if (!this.realJointIndexMap.has(link.id)) {
                 const connectedJoint = link.joints.find((j) => j instanceof RealJoint)!;
-                this.realJointIndexMap.set(link.id, joints.indexOf(connectedJoint));
+                // Match by id: the per-timestep joint arrays hold copies, so
+                // an identity indexOf against the link's original joints
+                // finds nothing and the index comes back -1.
+                this.realJointIndexMap.set(
+                  link.id,
+                  joints.findIndex((j) => j.id === connectedJoint.id)
+                );
                 const prisJoint = link.joints.find(jt => jt instanceof PrisJoint) as PrisJoint;
                 this.desiredAngleMap.set(connectedJoint.id, prisJoint.angle_rad);
               }
